requirements.txt
# Vector DB
qdrant-client==1.17.1

# Embeddings + Azure OpenAI
openai>=1.30.0

# Ingestion / parsing
unstructured==0.16.11
unstructured-inference==0.7.36
pymupdf==1.25.5
pdfplumber==0.11.4
playwright==1.50.0
beautifulsoup4==4.12.3
requests==2.32.3
pyyaml==6.0.2

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1